backend: split into backend_{host,cuda,hip, ...}.h
