+++ include/gtensor/backend_hip.h
@@ -1,0 +1,195 @@
+
+#ifndef GTENSOR_BACKEND_HIP_H
+#define GTENSOR_BACKEND_HIP_H
+
+#include "pointer_traits.h"
+
+#include <hip/hip_runtime.h>
+
+// ======================================================================
+// gt::backend::hip
+
+namespace gt
+{
+namespace backend
+{
+
+namespace hip
+{
+
+namespace detail
+{
+
+template <typename S_src, typename S_to>
+struct copy;
+
+template <>
+struct copy<space::device, space::device>
+{
+  template <typename T>
+  static void run(const T* src, T* dst, size_type count)
+  {
+    gtGpuCheck(hipMemcpy(dst, src, sizeof(T) * count, hipMemcpyDeviceToDevice));
+  }
+};
+
+template <>
+struct copy<space::device, space::host>
+{
+  template <typename T>
+  static void run(const T* src, T* dst, size_type count)
+  {
+    gtGpuCheck(hipMemcpy(dst, src, sizeof(T) * count, hipMemcpyHostToDevice));
+  }
+};
+
+template <>
+struct copy<space::host, space::device>
+{
+  template <typename T>
+  static void run(const T* src, T* dst, size_type count)
+  {
+    gtGpuCheck(hipMemcpy(dst, src, sizeof(T) * count, hipMemcpyHostToDevice));
+  }
+};
+
+template <>
+struct copy<space::host, space::host>
+{
+  template <typename T>
+  static void run(const T* src, T* dst, size_type count)
+  {
+    gtGpuCheck(hipMemcpy(dst, src, sizeof(T) * count, hipMemcpyHostToHost));
+  }
+};
+
+} // namespace detail
+
+template <typename S_src, typename S_to, typename T>
+inline void copy(const T* src, T* dst, gt::size_type count)
+{
+  return detail::copy<S_src, S_to>::run(src, dst, count);
+}
+
+struct ops
+{
+  static void memset(void* dst, int value, gt::size_type nbytes)
+  {
+    gtGpuCheck(hipMemset(dst, value, nbytes));
+  }
+};
+
+namespace gallocator
+{
+struct device
+{
+  template <typename T>
+  static T* allocate(size_type n)
+  {
+    T* p;
+    gtGpuCheck(hipMalloc(&p, sizeof(T) * n));
+    return p;
+  }
+
+  template <typename T>
+  static void deallocate(T* p)
+  {
+    gtGpuCheck(hipFree(p));
+  }
+};
+
+struct managed
+{
+  template <typename T>
+  static T* allocate(size_t n)
+  {
+    T* p;
+    gtGpuCheck(hipMallocManaged(&p, sizeof(T) * n));
+    return p;
+  }
+
+  template <typename T>
+  static void deallocate(T* p)
+  {
+    gtGpuCheck(hipFree(p));
+  }
+};
+
+struct host
+{
+  template <typename T>
+  static T* allocate(size_type n)
+  {
+    T* p;
+    gtGpuCheck(hipHostMalloc(&p, sizeof(T) * n));
+    return p;
+  }
+
+  template <typename T>
+  static void deallocate(T* p)
+  {
+    gtGpuCheck(hipHostFree(p));
+  }
+};
+
+} // namespace gallocator
+
+template <typename T>
+using device_allocator =
+  wrap_allocator<T, typename gallocator::device, gt::space::device>;
+
+template <typename T>
+using host_allocator =
+  wrap_allocator<T, typename gallocator::host, gt::space::host>;
+
+inline void device_synchronize()
+{
+  gtGpuCheck(hipStreamSynchronize(0));
+}
+
+template <typename T>
+inline void device_copy_async_dd(const T* src, T* dst, gt::size_type count)
+{
+  gtGpuCheck(
+    hipMemcpyAsync(dst, src, sizeof(T) * count, hipMemcpyDeviceToDevice));
+}
+
+inline int device_get_count()
+{
+  int device_count;
+  gtGpuCheck(hipGetDeviceCount(&device_count));
+  return device_count;
+}
+
+inline void device_set(int device_id)
+{
+  gtGpuCheck(hipSetDevice(device_id));
+}
+
+inline int device_get()
+{
+  int device_id;
+  gtGpuCheck(hipGetDevice(&device_id));
+  return device_id;
+}
+
+inline uint32_t device_get_vendor_id(int device_id)
+{
+  hipDeviceProp_t prop;
+  uint32_t packed = 0;
+
+  gtGpuCheck(hipGetDeviceProperties(&prop, device_id));
+
+  packed |= (0x000000FF & ((uint32_t)prop.pciDeviceID));
+  packed |= (0x0000FF00 & (((uint32_t)prop.pciBusID) << 8));
+  packed |= (0xFFFF0000 & (((uint32_t)prop.pciDomainID) << 16));
+
+  return packed;
+}
+
+} // namespace hip
+
+} // namespace backend
+} // namespace gt
+
+#endif // GTENSOR_BACKEND_HIP_H
