+++ include/gtensor/backend_sycl.h
@@ -1,0 +1,133 @@
+
+#ifndef GTENSOR_BACKEND_SYCL_H
+#define GTENSOR_BACKEND_SYCL_H
+
+#include "pointer_traits.h"
+#include "sycl_backend.h"
+
+#include <CL/sycl.hpp>
+
+// ======================================================================
+// gt::backend::sycl
+
+namespace gt
+{
+namespace backend
+{
+
+namespace sycl
+{
+
+template <typename S_src, typename S_to, typename T>
+inline void copy(const T* src, T* dst, gt::size_type count)
+{
+  cl::sycl::queue& q = gt::backend::sycl::get_queue();
+  q.memcpy(dst, src, sizeof(T) * count);
+  q.wait();
+}
+
+struct ops
+{
+  static void memset(void* dst, int value, gt::size_type nbytes)
+  {
+    cl::sycl::queue& q = gt::backend::sycl::get_queue();
+    q.memset(dst, value, nbytes);
+  }
+};
+
+namespace gallocator
+{
+struct device
+{
+  template <typename T>
+  static T* allocate(size_type n)
+  {
+    return cl::sycl::malloc_shared<T>(n, gt::backend::sycl::get_queue());
+  }
+
+  template <typename T>
+  static void deallocate(T* p)
+  {
+    cl::sycl::free(p, gt::backend::sycl::get_queue());
+  }
+};
+
+struct managed
+{
+  template <typename T>
+  static T* allocate(size_t n)
+  {
+    return cl::sycl::malloc_shared<T>(n, gt::backend::sycl::get_queue());
+  }
+
+  template <typename T>
+  static void deallocate(T* p)
+  {
+    cl::sycl::free(p, gt::backend::sycl::get_queue());
+  }
+};
+
+// The host allocation type in SYCL allows device code to directly access
+// the data. This is generally not necessary or effecient for gtensor, so
+// we opt for the same implementation as for the HOST device below.
+
+struct host
+{
+  template <typename T>
+  static T* allocate(size_type n)
+  {
+    T* p = static_cast<T*>(malloc(sizeof(T) * n));
+    if (!p) {
+      std::cerr << "host allocate failed" << std::endl;
+      std::abort();
+    }
+    return p;
+  }
+
+  template <typename T>
+  static void deallocate(T* p)
+  {
+    free(p);
+  }
+};
+
+// template <typename T>
+// struct host
+// {
+//   static T* allocate( : size_type count)
+//   {
+//     return cl::sycl::malloc_host<T>(count, gt::backend::sycl::get_queue());
+//   }
+
+//   static void deallocate(T* p)
+//   {
+//     cl::sycl::free(p, gt::backend::sycl::get_queue());
+//   }
+// };
+
+} // namespace gallocator
+
+template <typename T>
+using device_allocator =
+  wrap_allocator<T, typename gallocator::device, gt::space::device>;
+
+template <typename T>
+using host_allocator =
+  wrap_allocator<T, typename gallocator::host, gt::space::host>;
+
+inline void device_synchronize()
+{
+  gt::backend::sycl::get_queue().wait();
+}
+
+template <typename T>
+inline void device_copy_async_dd(const T* src, T* dst, gt::size_type count)
+{
+  cl::sycl::queue& q = gt::backend::sycl::get_queue();
+  q.memcpy(dst, src, sizeof(T) * count);
+}
+
+} // namespace sycl
+} // namespace backend
+
+#endif GTENSOR_BACKEND_SYCL_H
